# WIP: Experiment. Further optimize Obb::HasOverlap() using array().abs() instead of `for` loop.