--- a/geometry/benchmarking/obb_benchmark/BUILD.bazel
+++ b/geometry/benchmarking/obb_benchmark/BUILD.bazel
@@ -1,0 +1,33 @@
+# -*- python -*-
+
+load(
+    "@drake//tools/skylark:drake_cc.bzl",
+    "drake_cc_binary",
+    "drake_cc_library",
+    "drake_cc_test",
+)
+load("//tools/lint:lint.bzl", "add_lint_tests")
+
+drake_cc_test(
+    name = "obb_bench",
+    srcs = ["obb_bench.cc"],
+    linkstatic = True,
+    deps = [
+        "//common:essential",
+        "//geometry/proximity:obb",
+        "@googlebenchmark//:benchmark",
+    ],
+)
+
+sh_test(
+    name = "record_results",
+    size = "small",
+    srcs = ["record_results.sh"],
+    data = [
+        ":obb_bench",
+        "//tools/performance:record_results",
+    ],
+    tags = ["no_valgrind_tools"],
+)
+
+add_lint_tests()