--- a/geometry/benchmarking/obb_benchmark/conduct_experiment
+++ b/geometry/benchmarking/obb_benchmark/conduct_experiment
@@ -1,0 +1,19 @@
+#!/bin/bash
+# Execute the complete set of steps for a well-controlled benchmark experiment.
+
+set -e -u -o pipefail
+
+TARGET=//geometry/benchmarking/obb_benchmark:record_results
+ME=$(python3 -c 'import os; print(os.path.realpath("'"$0"'"))')
+HERE=$(dirname $ME)
+
+OUTPUT_DIR="$1"
+shift
+
+cd "$HERE"/../../..
+
+./tools/performance/benchmark_tool conduct_experiment \
+    -b=--compilation_mode=opt \
+    -b=--copt=-g \
+    -b=--cxxopt=-march=broadwell \
+    $TARGET "$OUTPUT_DIR" -- "$@"