--- a/geometry/benchmarking/obb_benchmark/copy_results_to
+++ b/geometry/benchmarking/obb_benchmark/copy_results_to
@@ -1,0 +1,12 @@
+#!/bin/bash
+# Copy benchmark results to a user-chosen directory.
+
+set -e -u -o pipefail
+
+TARGET=//geometry/benchmarking/obb_benchmark:record_results
+ME=$(python3 -c 'import os; print(os.path.realpath("'"$0"'"))')
+HERE=$(dirname $ME)
+
+cd "$HERE"/../../..
+
+./tools/performance/benchmark_tool copy_results $TARGET "$1"