--- a/geometry/benchmarking/obb_benchmark/record_results.sh
+++ b/geometry/benchmarking/obb_benchmark/record_results.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+# Collect context information for a benchmark experiment.
+
+${TEST_SRCDIR}/drake/tools/performance/record_results \
+    ${TEST_SRCDIR}/drake/geometry/benchmarking/obb_benchmark/obb_bench \
+    "$@"