(Rebase on ObbHasOverlapBenchMark #14621)
(Branch out of ObbHasOverlapAVXTransformOnly #14626)

Use array().abs() + kEpsilon instead of for{for{abs() + kEpsilon}}.

It can improve obb_benchmark slightly to:
```
$ tail -n12 ~/project/ObbBench/2021-02-05_1510_ObbAVXTransformArray/summary.txt 
-------------------------------------------------------------------------------------
Benchmark                                           Time             CPU   Iterations
-------------------------------------------------------------------------------------
BM_ObbHasOverlap_31_Cases_mean                   2922 ns         2922 ns            9
BM_ObbHasOverlap_31_Cases_median                 2919 ns         2919 ns            9
BM_ObbHasOverlap_31_Cases_stddev                 5.26 ns         5.25 ns            9
BM_ObbHasOverlap_16_OverlapCases_mean            1598 ns         1598 ns            9
BM_ObbHasOverlap_16_OverlapCases_median          1598 ns         1598 ns            9
BM_ObbHasOverlap_16_OverlapCases_stddev          1.19 ns         1.20 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_mean         1263 ns         1263 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_median       1264 ns         1263 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_stddev       1.90 ns         1.90 ns            9
```